+++ tokenauthy.services.yml
@@ -1,23 +1,18 @@
 services:
-  tokenauthy.authentication.auth_token:
-    class: Drupal\tokenauthy\Authentication\Provider\AuthToken
+  tokenauthy.middleware:
+    class: Drupal\tokenauthy\Stackmiddleware\AuthTokenMiddleware
     arguments:
       - '@session'
+      - '@session_configuration'
       - '@entity_type.manager'
       - '@tokenauthy'
     tags:
-      - { name: authentication_provider, provider_id: 'auth_token', priority: 50, global: TRUE }
+      - { name: http_middleware, priority: 20 }
   tokenauthy.page_cache_request_policy.by_pass_auth_token:
     class: Drupal\tokenauthy\PageCache\AllowAuthTokenRequests
     public: false
     tags:
       - { name: page_cache_request_policy }
-  tokenauthy.event_subscriber:
-    class: Drupal\tokenauthy\EventSubscriber\TokenAuthyEventSubscriber
-    arguments:
-      - '@session'
-    tags:
-      - { name: event_subscriber }
   tokenauthy:
     class: Drupal\tokenauthy\Services\TokenAuthy
     arguments:
